requirements.txt
Flask==2.3.2
Jinja2>=3.1.2
Werkzeug>=2.3.3
python-dotenv==0.21.0
gunicorn==20.1.0
Flask-Cors==3.0.10
ushlex==0.99.1
requests==2.30.0
argon2-cffi==21.3.0